# DimensionPredicateFilter: Skip unnecessary dictionary lookup.