Before:

```
Benchmark                                      (cardinality)  Mode  Cnt        Score       Error  Units
DimensionPredicateFilterBenchmark.matchIsEven           1000  avgt   10      836.358 ±    48.417  us/op
DimensionPredicateFilterBenchmark.matchIsEven         100000  avgt   10   120199.978 ±  5266.771  us/op
DimensionPredicateFilterBenchmark.matchIsEven        1000000  avgt   10  1224461.621 ± 47246.838  us/op
```

After:

```
Benchmark                                      (cardinality)  Mode  Cnt       Score       Error  Units
DimensionPredicateFilterBenchmark.matchIsEven           1000  avgt   10     349.276 ±    17.233  us/op
DimensionPredicateFilterBenchmark.matchIsEven         100000  avgt   10   35653.795 ±  1039.967  us/op
DimensionPredicateFilterBenchmark.matchIsEven        1000000  avgt   10  329050.110 ± 15731.258  us/op
```