@xvrl 2016-04-18
@gianm was most of the time spend in doing bitmap lookups, or creating predicates?