@gianm 2016-04-18
@xvrl didn't check that explicitly, although I assume mostly it was in the bitmap lookups, as I only removed one predicate creation per call to `getBitmapIndex` and don't think that would really be noticeable.